+++ client/pages/forecast.tsx
@@ -3,24 +3,14 @@ import type { GetServerSidePropsResult, NextPage } from 'next'
 import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { NextSeo } from 'next-seo'
-import { ColumnProps } from 'simple-react-ui-kit'
 
-import { API, ApiModel } from '@/api'
+import { API } from '@/api'
 import { setLocale } from '@/api/applicationSlice'
 import { wrapper } from '@/api/store'
 import AppLayout from '@/components/app-layout'
-import WeatherIcon from '@/components/weather-icon'
 import WidgetForecastTable from '@/components/widget-forecast-table'
-import styles from '@/components/widget-forecast-table/styles.module.sass'
 import WidgetMeteogram from '@/components/widget-meteogram'
-import WindDirectionIcon from '@/components/wind-direction-icon'
 import { POLING_INTERVAL_FORECAST } from '@/pages/_app'
-import { getWeatherI18nKey } from '@/tools/conditions'
-import { formatDate } from '@/tools/date'
-import { round } from '@/tools/helpers'
-import { convertHpaToMmHg, getCloudinessColor, getTemperatureColor } from '@/tools/weather'
-import ComparisonIcon from '@/ui/comparison-icon'
-
 type IndexPageProps = object
 
 const IndexPage: NextPage<IndexPageProps> = () => {
@@ -33,84 +23,6 @@ const IndexPage: NextPage<IndexPageProps> = () => {
     const { data: forecastDaily, isLoading: dailyLoading } = API.useGetForecastQuery('daily', {
         pollingInterval: POLING_INTERVAL_FORECAST
     })
-
-    const tableColumnsDaily: ColumnProps<ApiModel.Weather>[] = [
-        {
-            header: t('date'),
-            accessor: 'date',
-            className: styles.cellDate,
-            isSortable: true,
-            formatter: (date) => formatDate(date as string, 'dd, MMM D')
-        },
-        {
-            header: t('weather'),
-            accessor: 'weatherId',
-            className: styles.cellCondition,
-            formatter: (weatherId) => (
-                <>
-                    <WeatherIcon weatherId={weatherId as number} />
-                    {t(getWeatherI18nKey(weatherId || ''))}
-                </>
-            )
-        },
-        {
-            header: t('temperature-short'),
-            accessor: 'temperature',
-            className: styles.cellTemperature,
-            isSortable: true,
-            background: (temperature) => getTemperatureColor(temperature),
-            formatter: (temperature) => <>{round(Number(temperature), 1)} °C</>
-        },
-        {
-            header: t('clouds'),
-            accessor: 'clouds',
-            className: styles.cellClouds,
-            isSortable: true,
-            background: (clouds) => getCloudinessColor(clouds),
-            formatter: (clouds) => <>{clouds}%</>
-        },
-        {
-            header: t('pressure'),
-            accessor: 'pressure',
-            className: styles.cellPressure,
-            isSortable: true,
-            formatter: (pressure, data, i) => (
-                <>
-                    <span>{convertHpaToMmHg(pressure)}</span>
-                    <ComparisonIcon
-                        currentValue={pressure}
-                        previousValue={data[i - 1]?.pressure}
-                    />
-                </>
-            )
-        },
-        {
-            header: t('precipitation'),
-            accessor: 'precipitation',
-            className: styles.cellClouds,
-            isSortable: true,
-            formatter: (precipitation) =>
-                precipitation ? (
-                    <>
-                        {precipitation} {t('millimeters')}
-                    </>
-                ) : (
-                    ''
-                )
-        },
-        {
-            header: t('wind'),
-            accessor: 'windSpeed',
-            className: styles.cellWind,
-            isSortable: true,
-            formatter: (windSpeed, data, i) => (
-                <>
-                    <WindDirectionIcon direction={Number(data[i]?.windDeg)} />
-                    {round(Number(windSpeed), 1)} {t('meters-per-second')}
-                </>
-            )
-        }
-    ]
 
     return (
         <AppLayout>
@@ -144,8 +56,8 @@ const IndexPage: NextPage<IndexPageProps> = () => {
                 <WidgetForecastTable
                     title={t('weather-forecast-by-days')}
                     loading={dailyLoading}
-                    columns={tableColumnsDaily}
                     data={forecastDaily}
+                    columnsPreset={['date', 'weather', 'temperature', 'clouds', 'pressure', 'precipitation', 'wind']}
                     defaultSort={{ key: 'date', direction: 'asc' }}
                     fullWidth={true}
                 />